requirements.txt
Flask==2.3.2
Pillow==10.0.0
reportlab==4.0.0
numpy==1.26.0
